backend/requirements.txt
flask
flask_cors
flask_login
pymongo
google.generativeai
dotenv
PyPDF4
flask_session
gridfs
